restructure crate
